refactor(lotto): inputView에서의 의존하고 있는 객체 제거 리팩토링
